feat: RoomService 분리
