+++ src/main/resources/templates/home.hbs
@@ -5,6 +5,11 @@
     <title>체스</title>
     <script type="text/javascript" src="/src/home.js"></script>
 </head>
+<style>
+    body {
+        background-color:darkseagreen;
+    }
+</style>
 <body>
 <div align="center">
     <div>
