+++ src/main/resources/templates/result.hbs
@@ -1,13 +1,16 @@
 <!DOCTYPE html>
 <html lang="en">
+<style>
+    body {
+        background-color:darkseagreen;
+    }
+</style>
 <head>
     <meta charset="UTF-8">
     <title>게임 결과</title>
     <script>
         function calculate(score) {
-
         }
-
     </script>
 </head>
 <body>
@@ -26,6 +29,6 @@
 <p> 우승자 : {{winner}} </p>
 {{/result}}
 <br>
-<a href="/">홈으로 갈게~ㅎ</a>
+<a href="/">메인 화면으로 이동</a>
 </body>
 </html>
